parse_parenthesizedのテストを整理しました
